+++ src/realm/object-store/c_api/types.hpp
@@ -37,6 +37,10 @@ struct UnexpectedPrimaryKeyException : std::logic_error {
     using std::logic_error::logic_error;
 };
 
+struct DuplicatePrimaryKeyException : std::logic_error {
+    using std::logic_error::logic_error;
+};
+
 struct InvalidPropertyKeyException : std::logic_error {
     using std::logic_error::logic_error;
 };